Various dicom fixes
Summary:
Make the dicom parser more tolerant of corruption. Change dicom.ReadDataSet and its variant to return non-nil dataset even if parsing fails. The dataset will contain whatever elements the parser managed to salvage. The decoder and the parser was extended to skip over unparsable parts.

Sort the the tags in DIMSE response messages. dcmtk is not happy otherwise.

Extend dicom-server to serve files that contain parse errors, with "*" mark in the dashboard.

Extend dicom server to save the dicom attributes in a checkpoint file, so that it doesn't need to fully scan all the dicom  files when it restarts.

Limit the file-read parallelism of dicom server. Before this change the server could read all the files at once, and it could cause the s3-read requests to time out.

Reviewers: jbredno

Reviewed By: jbredno

Differential Revision: https://phabricator.grailbio.com/D11388

fbshipit-source-id: 1afbeb6